Implemented API rate limiting

+++ src/main/java/com/asr/springboot_mongodb/controller/PersonController.java
@@ -1,8 +1,10 @@
 package com.asr.springboot_mongodb.controller;
 
 import com.asr.springboot_mongodb.collection.Person;
+import com.asr.springboot_mongodb.exceptions.ApiLimitExceedException;
+import com.asr.springboot_mongodb.ratelimiter.RateLimiter;
 import com.asr.springboot_mongodb.service.PersonService;
-import io.swagger.models.auth.In;
+import lombok.SneakyThrows;
 import org.bson.Document;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -17,28 +19,38 @@ import java.util.List;
 public class PersonController {
     @Autowired
     private PersonService personService;
+    @Autowired
+    private RateLimiter rateLimiter;
 
     @PostMapping
+    @SneakyThrows
     public String save(@RequestBody Person person){
         return personService.save(person);
     }
 
     @GetMapping
-    public List<Person> getPersonNameStartsWith(@RequestParam("name") String name){
+    @SneakyThrows
+    public List<Person> getPersonNameStartsWith(@RequestParam("name") String name) {
+        if( !rateLimiter.tryGet() ) throw new ApiLimitExceedException("Too many requests");
         return personService.getPersonNameStartsWith(name);
     }
 
     @DeleteMapping
-    public void delete(@PathVariable String id){
+    @SneakyThrows
+    public void delete(@PathVariable String id) throws Exception {
+        if( !rateLimiter.tryGet() ) throw new ApiLimitExceedException("Too many requests");
         personService.delete(id);
     }
 
     @GetMapping("/age")
+    @SneakyThrows
     public List<Person> getByPersonAge(@RequestParam Integer minAge, @RequestParam Integer maxAge){
+        if( !rateLimiter.tryGet() ) throw new ApiLimitExceedException("Too many requests");
         return personService.getAgeByMinAgeMaxAge(minAge,maxAge);
     }
 
     @GetMapping("/search")
+    @SneakyThrows
     public Page<Person> searchPerson(@RequestParam(required = false) String name,
                                      @RequestParam(required = false) String minAge,
                                      @RequestParam(required = false) String maxAge,
@@ -46,17 +58,22 @@ public class PersonController {
                                      @RequestParam(defaultValue = "0") Integer page,
                                      @RequestParam(defaultValue = "5") Integer size
     ) {
+        if( !rateLimiter.tryGet() ) throw new ApiLimitExceedException("Too many requests");
         Pageable pageable = PageRequest.of(page,size);
         return personService.search(name, minAge,maxAge,city,page,size,pageable);
     }
 
     @GetMapping("/oldestPerson")
+    @SneakyThrows
     public List<Document> getOldestPerson(){
+        if( !rateLimiter.tryGet() ) throw new ApiLimitExceedException("Too many requests");
         return personService.getOldestPersonByCity();
     }
 
     @GetMapping("/populationByCity")
+    @SneakyThrows
     public List<Document> getPopulationByCity(){
+        if( !rateLimiter.tryGet() ) throw new ApiLimitExceedException("Too many requests");
         return personService.getPopulationByCity();
     }
 }
